backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pypdf==4.0.2
pdf2docx==0.5.8
python-docx==1.1.0
Pillow==10.2.0
prometheus-client==0.19.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.4
httpx==0.26.0
